Pulling out a base class to avoid the duplication.
