+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/References/UnSetAttributeCommand.cs
@@ -1,60 +1,16 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.References
 {
-    internal class UnSetAttributeCommand : IProjectSystemUpdateReferenceOperation
+    internal class UnsetAttributeCommand : BaseSetAttributeCommand
     {
-        private readonly ConfiguredProject _selectedConfiguredProject;
-        private readonly string _itemSpecification;
-        private readonly AbstractReferenceHandler _referenceHandler;
+        public UnsetAttributeCommand(AbstractReferenceHandler abstractReferenceHandler, ConfiguredProject selectedConfiguredProject, string itemSpecification)
+        : base(abstractReferenceHandler, selectedConfiguredProject, itemSpecification) {
 
-        public UnSetAttributeCommand(AbstractReferenceHandler abstractReferenceHandler, ConfiguredProject selectedConfiguredProject, string itemSpecification)
-        {
-            _referenceHandler = abstractReferenceHandler;
-            _selectedConfiguredProject = selectedConfiguredProject;
-            _itemSpecification = itemSpecification;
-        }
-
-        public async Task<bool> ApplyAsync(CancellationToken cancellationToken)
-        {
-            IProjectItem? item = await GetProjectItemAsync();
-
-            if (item == null)
-            {
-                return false;
-            }
-
-            await item.Metadata.SetPropertyValueAsync(ProjectReference.TreatAsUsedProperty, PropertySerializer.SimpleTypes.ToString(false), null);
-
-            return true;
-        }
-
-        public async Task<bool> RevertAsync(CancellationToken cancellationToken)
-        {
-            IProjectItem? item = await GetProjectItemAsync();
-
-            if (item == null)
-            {
-                return false;
-            }
-
-            await item.Metadata.SetPropertyValueAsync(ProjectReference.TreatAsUsedProperty, PropertySerializer.SimpleTypes.ToString(true), null);
-
-            return true;
-        }
-
-        private async Task<IProjectItem?> GetProjectItemAsync()
-        {
-            var projectItems = await _referenceHandler.GetUnresolvedReferencesAsync(_selectedConfiguredProject);
-
-            var item = projectItems
-                .FirstOrDefault(c => c.EvaluatedInclude == _itemSpecification);
-            return item;
+            _unsetValue = PropertySerializer.SimpleTypes.ToString(true);
+            _setValue = PropertySerializer.SimpleTypes.ToString(false);
         }
     }
 }
